Hej Lærke. Jeg har ryddet op i mainfilen. Dvs. at level 1 og 2 nu fremgår i en selvstændig fil.
